fix(trafficParser): handle trash loot dropping before first boss kill
Previously, a trash item could end up in the wrong raid if it dropped
before the first boss kill in a raid. If the trash loot doesn't match
the current raid, it is added in a pending state until the next raid
boss, at which point it is added to the trash list.