feat(trafficParser): parse from LUA instead of JSON
Previously the JSON was exported from the addon, however the addon
doesn't officially support a JSON addon. Instead of exporting JSON from
the addon, the LUA can be pasted in directly.

Fengari is used to execute lua, the LUA is used to return the CEPGP
variable. The tables are then looked up.